requirements.txt
Flask
folium
mysql-connector-python
